Add files via upload
**Delve into the Sonic Labyrinth: My Latest EP, FLUX**

Prepare to be captivated by my latest EP, FLUX, a mesmerizing fusion of captivating melodies and pulsating rhythms. Embark on a sonic odyssey as each track unveils a unique sonic tapestry, transporting you to a realm of emotions, from the electrifying energy of "Surge" to the introspective tranquility of "Serenity."

**Immerse Yourself in FLUX:**

- Plunge into the aural experience: FLUX on flux.nctrnm.com

- Witness the visual spectacle: FLUXV on fluxv.nctrnm.com

- Enhance your FLUX journey with official merchandise: FLUX Merchandise on shop.nctrnm.com

FLUX transcends the boundaries of an EP, offering an immersive exploration of my artistry. Each track is meticulously crafted to evoke a distinct emotion, leaving you breathless and yearning for more.

**This is my personal invitation to experience the sonic journey of FLUX**

Signed-off-by: Matthew McGilvery <48932877+nctrnm@users.noreply.github.com>